+++ TimetableManagementSystem/src/main/java/tj132/Tana/Personal_Schedule.java
@@ -1,4 +1,4 @@
-package teacher;
+package tj132.Tana;
 import java.time.LocalDateTime;
 
 public class Personal_Schedule {
