+++ TimetableManagementSystem/src/main/java/tj132/Tana/Personal_ScheduleTest.java
@@ -1,4 +1,4 @@
-package teacher;
+package tj132.Tana;
 
 import static org.junit.jupiter.api.Assertions.*;
