+++ TimetableManagementSystem/src/main/java/tj132/Tana/personalmgnt.java
@@ -1,4 +1,4 @@
-package teacher;
+package tj132.Tana;
 
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
