+++ TimetableManagementSystem/src/main/java/tj132/Tana/tt_admin.java
@@ -1,4 +1,4 @@
-package admin_mgmt;
+package tj132.Tana;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -17,8 +17,8 @@ import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Filters;
 import com.mongodb.client.model.Updates;
 
-import teacher.Personal_Schedule;
-import teacher.personalmgnt.contact;
+import tj132.Tana.Personal_Schedule;
+import tj132.Tana.personalmgnt.contact;
 
 public class tt_admin {
